# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/companies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtener lista de compañías con filtros y paginación",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Obtener lista de compañías",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre de la empresa",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nombre legal de la empresa",
                        "name": "legal_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Identificador fiscal",
                        "name": "tax_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Estado activo/inactivo",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Tamaño de página",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campo para ordenar",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "Dirección (asc o desc)",
                        "name": "sort_direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.PaginatedResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crear una nueva compañia",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Crear una nueva compañia",
                "parameters": [
                    {
                        "description": "Company information",
                        "name": "company",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/companies/addresses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtener todas las direcciones de la compañia autenticada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "addresses"
                ],
                "summary": "Obtener todas las direcciones de la compañia autenticada",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyAddressResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Agregar una nueva dirección para la compañia autenticada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "addresses"
                ],
                "summary": "Agregar una nueva dirección para la compañia autenticada",
                "parameters": [
                    {
                        "description": "Address information",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyAddressDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Address added successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/companies/addresses/{address_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualizar una dirección para la compañia autenticada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "addresses"
                ],
                "summary": "Actualizar una dirección para la compañia autenticada",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address ID",
                        "name": "address_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Address information",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyAddressDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Eliminar una dirección de la compañia autenticada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "addresses"
                ],
                "summary": "Eliminar una dirección de la compañia autenticada",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address ID",
                        "name": "address_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/companies/deactivate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deactivate the authenticated company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Deactivate the authenticated company",
                "responses": {
                    "200": {
                        "description": "Company deactivated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/companies/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtener la información de principal de la compañia",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Obtener la información de principal de la compañia",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualizar la información de la compañia",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Actualizar la información de la compañia",
                "parameters": [
                    {
                        "description": "Company information",
                        "name": "company",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/companies/reactivate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reactivate the authenticated company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Reactivate the authenticated company",
                "responses": {
                    "200": {
                        "description": "Company reactivated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyAddressDTO": {
            "type": "object",
            "required": [
                "address_line1",
                "city",
                "state"
            ],
            "properties": {
                "address_line1": {
                    "description": "Primera línea de dirección",
                    "type": "string",
                    "example": "Calle 100 #15-20"
                },
                "address_line2": {
                    "description": "Segunda línea de dirección (opcional)",
                    "type": "string",
                    "example": "Edificio Centro Empresarial"
                },
                "city": {
                    "description": "Ciudad",
                    "type": "string",
                    "example": "Bogotá"
                },
                "is_main": {
                    "description": "Indica si es la dirección principal",
                    "type": "boolean",
                    "example": true
                },
                "latitude": {
                    "description": "Latitud de la ubicación",
                    "type": "number",
                    "example": 4.68
                },
                "longitude": {
                    "description": "Longitud de la ubicación",
                    "type": "number",
                    "example": -74.05
                },
                "postal_code": {
                    "description": "Código postal",
                    "type": "string",
                    "example": "110121"
                },
                "state": {
                    "description": "Estado o provincia",
                    "type": "string",
                    "example": "Cundinamarca"
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyCreateRequest": {
            "type": "object",
            "required": [
                "contact_email",
                "contact_phone",
                "contract_start_date",
                "delivery_rate",
                "legal_name",
                "main_address",
                "name",
                "tax_id"
            ],
            "properties": {
                "contact_email": {
                    "description": "Email de contacto principal de la empresa",
                    "type": "string",
                    "example": "contacto@expressdelivery.com"
                },
                "contact_phone": {
                    "description": "Teléfono de contacto principal",
                    "type": "string",
                    "example": "+573001234567"
                },
                "contract_details": {
                    "description": "Detalles del contrato en formato JSON",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.ContractDetailsDTO"
                        }
                    ]
                },
                "contract_end_date": {
                    "description": "Fecha de fin del contrato (opcional)",
                    "type": "string",
                    "format": "date-time"
                },
                "contract_start_date": {
                    "description": "Fecha de inicio del contrato",
                    "type": "string",
                    "format": "date-time"
                },
                "delivery_rate": {
                    "description": "Tarifa base de entrega",
                    "type": "number",
                    "example": 20.5
                },
                "legal_name": {
                    "description": "Nombre legal completo de la empresa",
                    "type": "string",
                    "example": "Express Delivery S.A.S"
                },
                "logo_url": {
                    "description": "URL del logo de la empresa (opcional)",
                    "type": "string",
                    "example": "https://www.example.com/logo.png"
                },
                "main_address": {
                    "description": "Dirección principal de la empresa",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.CompanyAddressDTO"
                        }
                    ]
                },
                "name": {
                    "description": "Nombre comercial de la empresa",
                    "type": "string",
                    "example": "Express Delivery Co."
                },
                "tax_id": {
                    "description": "Identificador fiscal de la empresa",
                    "type": "string",
                    "example": "900123456-7"
                },
                "website": {
                    "description": "Sitio web de la empresa (opcional)",
                    "type": "string",
                    "example": "https://www.expressdelivery.com"
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.PaginatedResponse": {
            "description": "Paginated data response with metadata about pagination",
            "type": "object",
            "properties": {
                "data": {
                    "description": "The actual data items (array of objects)"
                },
                "page": {
                    "description": "Current page number",
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "description": "Number of items per page",
                    "type": "integer",
                    "example": 20
                },
                "total_items": {
                    "description": "Total number of items across all pages",
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "description": "Total number of pages",
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIError"
                }
            }
        },
...
    }
}